Make the model more widget friendly.
Adding the model to a QListView etc would
result in exactly nothing being shown, that is
not ideal.

It is important to return QVariant() for roles < UserRole
since otherwise the view gets invalid size hints for example.

For more fun just randomly dump the first really user generated
role into the view, that gives at least something visual.
It needs a make-over though. But by default we could really
just pretty-format all data we have for the DisplayRole.