Implement authentication structure
EnginioIdentity classs is based class for all authentication calls,
it purpose is simple: get session token which would be used during
Enginio server communication.

EnginioAuthentication class is simple user/password authentication
method, which connects to Enginio server and query users there.

Current, approach is to login via an asynchronous call, but in future
it would be nice to have also a synchronous way to do it.