Fix crash in EnginioModel caused by EnginioClient destructor.
EnginioModel enginioChanged signal is connected to QueryChanged functor,
because changing the client can change completely the data owned by
the model, therefore the full restart is needed. EnginioClient destroy
signal set enginio property on the model to null causing enginioChanged
signal emission, which tried to do a new request with a null engine
pointer. This change adds the check.