Fix a memory leak in QML
The EnginioQmlReply needs to be deleted at some point. normally we would
connect finished signal to delete later slot, but the last one is not
accessible from QML and it would not be nice to force qml users to do
such low level operations. So for each EnginioQmlReply, on finished
signal, we transfer ownership from EnginioClient to JS environment.