Fix chunked upload chunk-size and test
The position of the QIODevice turns out to not work,
when it's read, sometimes the position advances too far.
This is only the case on tiny chunk-sizes,
as is the case in the test.

The test sets a small chunk size which means the image will
be uploaded in 8 parts.