Sanitize example dir structure
Because of the way the documentation is built
we can only have unique project names for the examples.
Since this change will break deployment of examples yet again,
the directory structure has been sanitized to have it clean
in the future.

This has the advantage that the examples show up sensibly in
Qt Creator as well - instead of two times the same name.