Introduce AttachedData concept in EnginioModelPrivate
Quite often we need to add internal data to _data. We can not do it
directly because such data would be visible to an user. AttachedData
is not meant to be used for all items in the model, but only for items
in a special state, for example during an update. Access to the state is
through an object id which is constant, so more reliable then row
number. That feature should fix _synced flag and some race conditions:

 - Appending two, independent items.
 - Appending an item and removing an other one