Initial websocket notifications experiment
Add EnginioBackendConnection which implements a minimally needed
WebSocket protocol to be able to read incoming messages from a
specified backend.

Add a test to set up a backend and test if we receive all the
notifications about the changes.
The backend has to be set up with the additional option:

settings: {
    websocket: true  # this enables websocket notifications
}

This currently only works with the staging server.