Fix package creation
We did not actually pack the libraries we modified into the package
but the ones found in the root build dir, thus it is a mystery to me
how the Linux packages worked without the fixed rpath libs...
Change the precedence so that we pick up everything in dist/build
and only fall back to the root build dir if the given path does not
exist.
Additionally preserve symlinks when packaging since on Mac the
frameworks have multiple symlinks to point to the correct version of
the library.

Change-Id: Id1a44c21eb8846ac2dfc3f27c60a82b7327f35fa
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@digia.com>