Remove old buildsystem hacks
I am not sure what the logic is, but qml_plugin.prf
adds _declarative to the library name unless we declare
the no_cxx_module. Copied from controls.

By doing that and loading the qml_plugin late we get the right
TARGET name and we get it modified the way the platform specifies.
This should fix iOS overriding the armv7 build with the lib for
the simulator which should have a different postfix (which
we cut off with our own magic).

Change-Id: I4b0bb39c2c068b78a9aeee0f0f2994941307d84b
Reviewed-by: Andras Becsi <andras.becsi@digia.com>