Doc: Avoid linking to collision pages
C++ class and QML type references for EnginioModel use a shared
doc-include file to document a number of property/method.

This commit splits the model-query.qdocinc content in two, so
that we can pass a correctly disambiguated link as an argument
from the documentation that uses the include. This way, we
avoid linking to collision pages.

Also moves the 'see also' links to the actual property/method
documentation, fixes the \return commands, and does some other
cleanup.

Change-Id: Ia6e178f0ee91625e62289134eada7fb1edc85f62
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@digia.com>