Fix debug-and-release build
Enginio is part of the Qt release and there is no need to do magic hacks
to install the qmldir file.
This was supposed to help running examples separately, but we assume
that the lib is shipped with Qt proper.

The problem was that the qmldir file would be simultaneously copied into
the same location twice which will fail on Windows.

Change-Id: I57dc18af361f4bdfc00c8c4a12bbe1f0986eb34f
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@theqtcompany.com>